+++ src/main/java/me/jcala/jmooc/entity/Chapter.java
@@ -15,7 +15,7 @@ public class Chapter implements Serializable{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;//id
+    private long id;//id
 
     @Column(nullable = false,length = 40)
     private String name;//章节名称
